Move exception specific stuff to exception.c
